backend/requirements.txt
Django>=4.1  # Versión base de Django
djangorestframework  # Para crear APIs REST
djangorestframework-simplejwt  # Para manejar la autenticación con JWT
psycopg2-binary  # Conector de PostgreSQL para Django
django-cors-headers  # Para habilitar CORS en Django
